+++ Assets/Examples/SaveLoadModule/Example1 WithZenjectInstaller/Script/Data/PlayerData.cs
@@ -3,20 +3,20 @@ namespace SaveLoadModule.Example1
     public class PlayerData
     {
         //private variable
-        [CustomSerializable] private float _hp = 100;
+        [CustomSerializable] private int _hp = 100;
         [CustomSerializable]
         private SkillData _skillData = new SkillData();
 
 
         //public variable
-        public float Hp => _hp;
+        public int Hp => _hp;
 
-        [CustomSerializable] public float Mp { get; private set; } = 50;
+        [CustomSerializable] public int Mp { get; private set; } = 50;
         
         
         //public method
-        public void SetHp(float hp) => _hp = hp;
+        public void SetHp(int hp) => _hp = hp;
 
-        public void SetMp(float mp) => Mp = mp;
+        public void SetMp(int mp) => Mp = mp;
     }
 }
